third_party/blink/web_tests/fast/forms/color/input-value-sanitization-color.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Tests for value sanitization algorithm.');

var input;

var fallbackValue = '"#000000"'

input = document.createElement('input');
input.type = 'color';
shouldBe('input.value', fallbackValue);

// Valid values: hex
shouldBe('input.value = ""; input.value', fallbackValue);
shouldBe('input.value = "#000000"; input.value', '"#000000"');
shouldBe('input.value = "#123456"; input.value', '"#123456"');
shouldBe('input.value = "#789abc"; input.value', '"#789abc"');
shouldBe('input.value = "#defABC"; input.value', '"#defabc"');
shouldBe('input.value = "#DEF012"; input.value', '"#def012"');

// Valid values: named colors
shouldBe('input.value = "black"; input.value', '"#000000"');
shouldBe('input.value = "blue"; input.value', '"#0000ff"');
shouldBe('input.value = "red"; input.value', '"#ff0000"');
shouldBe('input.value = "purple"; input.value', '"#800080"');
shouldBe('input.value = "green"; input.value', '"#008000"');
shouldBe('input.value = "cyan"; input.value', '"#00ffff"');
shouldBe('input.value = "yellow"; input.value', '"#ffff00"');
shouldBe('input.value = "white"; input.value', '"#ffffff"');
shouldBe('input.value = "White"; input.value', '"#ffffff"');
shouldBe('input.value = "WHITE"; input.value', '"#ffffff"');

// Valid values: shorthand hex, spaces, and rgb()
shouldBe('input.value = "#FFF"; input.value', '"#ffffff"');
shouldBe('input.value = " #ffffff"; input.value', '"#ffffff"');
shouldBe('input.value = "#ffffff "; input.value', '"#ffffff"');
shouldBe('input.value = "#00112233"; input.value', '"#001122"');
shouldBe('input.value = "rgb(0,0,0)"; input.value', '"#000000"');

// Invalid values
shouldBe('input.value = "000000"; input.value', fallbackValue);
shouldBe('input.value = "xxx-non-existent-color-name"; input.value', fallbackValue);
shouldBe('input.value = "transparent"; input.value', fallbackValue);
</script>
</body>
</html>
