third_party/blink/web_tests/fast/forms/color/input-value-sanitization-color-expected.txt

Tests for value sanitization algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.value is "#000000"
PASS input.value = ""; input.value is "#000000"
PASS input.value = "#000000"; input.value is "#000000"
PASS input.value = "#123456"; input.value is "#123456"
PASS input.value = "#789abc"; input.value is "#789abc"
PASS input.value = "#defABC"; input.value is "#defabc"
PASS input.value = "#DEF012"; input.value is "#def012"
PASS input.value = "black"; input.value is "#000000"
PASS input.value = "blue"; input.value is "#000000"
PASS input.value = "red"; input.value is "#000000"
PASS input.value = "purple"; input.value is "#000000"
PASS input.value = "green"; input.value is "#000000"
PASS input.value = "cyan"; input.value is "#000000"
PASS input.value = "yellow"; input.value is "#000000"
PASS input.value = "white"; input.value is "#000000"
PASS input.value = "White"; input.value is "#000000"
PASS input.value = "WHITE"; input.value is "#000000"
PASS input.value = "000000"; input.value is "#000000"
PASS input.value = "#FFF"; input.value is "#000000"
PASS input.value = " #ffffff"; input.value is "#000000"
PASS input.value = "#ffffff "; input.value is "#000000"
PASS input.value = "#00112233"; input.value is "#000000"
PASS input.value = "rgb(0,0,0)"; input.value is "#000000"
PASS input.value = "xxx-non-existent-color-name"; input.value is "#000000"
PASS input.value = "transparent"; input.value is "#000000"
PASS successfullyParsed is true

TEST COMPLETE
